make test dir, add conftest
