+++ .github/workflows/continuous-integration.yml
@@ -25,7 +25,7 @@ jobs:
         run: python -m pip install .[dev]
 
       - name: Run tests
-        run: pytest -v
+        run: pytest tests -v
 
       - name: Lint
         run: pre-commit run --all-files
